backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# PPT generation
python-pptx>=0.6.23

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.26.0

# Preview generation (optional, for LibreOffice-based conversion)
# python-libreoffice is not a package; use system libreoffice

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0